+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - create_header: Create standard dictionary header.
 
 ### Changed
+- get_status: Added call to create_header to use standard header for all calls.
 - list_nodes: Converted output to dictionary format and call data_out function to handle the output.
 - list_repos: Converted output to dictionary format and call data_out function to handle the output.
 - list_master: Converted output to dictionary format and call data_out function to handle the output.